Changed a cast (from guint to guint64) in a call to gtk_tree_store_set().  The data is actually 64-bit, and was being fetched that way - but not stored that way - leading to some very strange (and large) values.
Reviewed-on: https://code.wireshark.org/review/2468

+++ ui/gtk/lbm_stream_dlg.c
@@ -598,7 +598,7 @@ static gboolean lbmc_stream_dlg_tap_packet(void * tap_data, packet_info * pinfo,
             LBMC_STREAM_DLG_STORE_BYTES_COLUMN, (guint)0,
             LBMC_STREAM_DLG_STORE_FIRST_FRAME_COLUMN, (guint)0,
             LBMC_STREAM_DLG_STORE_LAST_FRAME_COLUMN, (guint)0,
-            LBMC_STREAM_DLG_STORE_CHANNEL_COLUMN, (guint)stream->channel,
+            LBMC_STREAM_DLG_STORE_CHANNEL_COLUMN, (guint64)stream->channel,
             LBMC_STREAM_DLG_STORE_SUBSTREAM_COLUMN, (guint)0,
             -1);
     }
